Move Dependency tracking functions to HttpRegion utils
