+++ src/plugins/core/metaData/refMetaDataHandler.ts
@@ -1,7 +1,6 @@
 import { log } from '../../../io';
 import * as models from '../../../models';
 import * as utils from '../../../utils';
-import { toEnvironmentKey } from '../../../utils';
 import { ImportProcessorContext } from './importMetaDataHandler';
 
 export function refMetaDataHandler(type: string, name: string | undefined, context: models.ParserContext): boolean {
@@ -36,7 +35,13 @@ class RefMetaAction {
     }
 
     if (reference) {
-      registerRegionDependent(context, reference.httpFile, reference.httpRegion, context.httpFile, context.httpRegion);
+      utils.registerRegionDependent(
+        context,
+        reference.httpFile,
+        reference.httpRegion,
+        context.httpFile,
+        context.httpRegion
+      );
       const envKey = utils.toEnvironmentKey(context.httpFile.activeEnvironment);
       log.trace('import variables', reference.httpRegion.variablesPerEnv[envKey]);
       Object.assign(context.variables, reference.httpRegion.variablesPerEnv[envKey]);
@@ -82,54 +87,4 @@ class RefMetaAction {
     }
     return undefined;
   }
-}
-
-export function registerRegionDependent(
-  context: ImportProcessorContext,
-  refFile: models.HttpFile,
-  refRegion: models.HttpRegion,
-  dependentFile: models.HttpFile,
-  dependentRegion: models.HttpRegion
-): void {
-  context.options.refDependencies ||= [];
-  let refDepEntry = context.options.refDependencies.find(e => e.refFile === refFile && e.refRegion === refRegion);
-  if (!refDepEntry) {
-    refDepEntry = { refFile, refRegion, dependents: [] };
-    context.options.refDependencies.push(refDepEntry);
-  }
-  const depEntry = refDepEntry.dependents.find(d => d.depFile === dependentFile && d.depRegion === dependentRegion);
-  if (!depEntry) {
-    refDepEntry.dependents.push({
-      depFile: dependentFile,
-      depRegion: dependentRegion,
-    });
-  }
-}
-
-function resetDependentRegionsWithVisitor(
-  context: ImportProcessorContext,
-  refFile: models.HttpFile,
-  refRegion: models.HttpRegion,
-  visitedDependents: Array<{ depFile: models.HttpFile; depRegion: models.HttpRegion }>
-): void {
-  const refDepEntry = context.options.refDependencies?.find(e => e.refFile === refFile && e.refRegion === refRegion);
-  if (!refDepEntry) return;
-  const unvisitedDependents = refDepEntry.dependents.filter(
-    d => !visitedDependents.find(v => v.depFile === d.depFile && v.depRegion === d.depRegion)
-  );
-  for (const { depFile, depRegion } of unvisitedDependents) {
-    delete depRegion.response;
-    delete depRegion.variablesPerEnv[toEnvironmentKey(depFile.activeEnvironment)];
-
-    visitedDependents.push({ depFile, depRegion });
-    resetDependentRegionsWithVisitor(context, depFile, depRegion, visitedDependents);
-  }
-}
-
-export function resetDependentRegions(
-  context: ImportProcessorContext,
-  refFile: models.HttpFile,
-  refRegion: models.HttpRegion
-): void {
-  resetDependentRegionsWithVisitor(context, refFile, refRegion, []);
 }
