+++ src/utils/httpRegionUtils.ts
@@ -1,5 +1,4 @@
 import * as models from '../models';
-import { resetDependentRegions } from '../plugins/core/metaData/refMetaDataHandler';
 import { toEnvironmentKey } from './environmentUtils';
 import { report } from './logUtils';
 import { cloneResponse } from './requestUtils';
@@ -78,6 +77,7 @@ export async function processHttpRegionActions(
     resetDependentRegions(context, context.httpFile, context.httpRegion);
   }
 }
+
 function initRegionScopedVariables(context: models.ProcessorContext) {
   const env = toEnvironmentKey(context.httpFile.activeEnvironment);
 
@@ -147,6 +147,72 @@ export function toProcessedHttpRegion(context: models.ProcessorContext): models.
     },
     response: context.httpRegion.response && cloneResponse(context.httpRegion.response),
   };
+}
+
+interface ImportRegionDependentsEntry {
+  refFile: models.HttpFile;
+  refRegion: models.HttpRegion;
+  dependents: Array<{
+    depFile: models.HttpFile;
+    depRegion: models.HttpRegion;
+  }>;
+}
+
+interface HttpRegionDepedenciesContext extends models.ProcessorContext {
+  options: {
+    refDependencies?: Array<ImportRegionDependentsEntry>;
+  };
+}
+
+export function registerRegionDependent(
+  context: models.ProcessorContext,
+  refFile: models.HttpFile,
+  refRegion: models.HttpRegion,
+  dependentFile: models.HttpFile,
+  dependentRegion: models.HttpRegion
+): void {
+  const depContext = context as HttpRegionDepedenciesContext;
+  depContext.options.refDependencies ||= [];
+  let refDepEntry = depContext.options.refDependencies.find(e => e.refFile === refFile && e.refRegion === refRegion);
+  if (!refDepEntry) {
+    refDepEntry = { refFile, refRegion, dependents: [] };
+    depContext.options.refDependencies.push(refDepEntry);
+  }
+  const depEntry = refDepEntry.dependents.find(d => d.depFile === dependentFile && d.depRegion === dependentRegion);
+  if (!depEntry) {
+    refDepEntry.dependents.push({
+      depFile: dependentFile,
+      depRegion: dependentRegion,
+    });
+  }
+}
+
+function resetDependentRegionsWithVisitor(
+  context: HttpRegionDepedenciesContext,
+  refFile: models.HttpFile,
+  refRegion: models.HttpRegion,
+  visitedDependents: Array<{ depFile: models.HttpFile; depRegion: models.HttpRegion }>
+): void {
+  const refDepEntry = context.options.refDependencies?.find(e => e.refFile === refFile && e.refRegion === refRegion);
+  if (!refDepEntry) return;
+  const unvisitedDependents = refDepEntry.dependents.filter(
+    d => !visitedDependents.find(v => v.depFile === d.depFile && v.depRegion === d.depRegion)
+  );
+  for (const { depFile, depRegion } of unvisitedDependents) {
+    delete depRegion.response;
+    delete depRegion.variablesPerEnv[toEnvironmentKey(depFile.activeEnvironment)];
+
+    visitedDependents.push({ depFile, depRegion });
+    resetDependentRegionsWithVisitor(context, depFile, depRegion, visitedDependents);
+  }
+}
+
+export function resetDependentRegions(
+  context: models.ProcessorContext,
+  refFile: models.HttpFile,
+  refRegion: models.HttpRegion
+): void {
+  resetDependentRegionsWithVisitor(context, refFile, refRegion, []);
 }
 
 export function isGlobalHttpRegion(httpRegion: models.HttpRegion): boolean {
